Add update purchase status endpoint

+++ teacup-loan-purchases/src/main/java/com/example/teacuploanpurchases/entity/PurchaseRepository.java
@@ -2,6 +2,6 @@ package com.example.teacuploanpurchases.entity;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PurchaseRepository extends CrudRepository<Purchase, Long> {
+public interface PurchaseRepository extends CrudRepository<Purchase, Integer> {
 
 }
